started dash video playing

+++ transcodeffmpeg.cpp
@@ -170,6 +170,7 @@ bool TranscodeFFmpeg::set_input(const char* time, const char* input, bool verbos
     long duration = 0;
     copy_audio = false;
     copy_video = false;
+    seekPossible = true;
 
     if (infoFile) {
         char buffer[1024];
@@ -185,7 +186,13 @@ bool TranscodeFFmpeg::set_input(const char* time, const char* input, bool verbos
                 }
             } else {
                 // duration
-                duration = std::max(strtol(line, (char **) NULL, 10), duration);
+                if (strncmp(line, "N/A", 3) == 0) {
+                    // e.g. for mpeg-dash. Set to 2 hours
+                    duration = 2 * 60 * 60;
+                    seekPossible = false;
+                } else {
+                    duration = std::max(strtol(line, (char **) NULL, 10), duration);
+                }
             }
 
             idx++;
@@ -264,7 +271,11 @@ bool TranscodeFFmpeg::fork_ffmpeg(long start_at_ms) {
             cmdline += "-hide_banner -loglevel warning ";
         }
 
-        cmdline += "-re -ss " + std::string(ms_to_ffmpeg_time(start_at_ms)) + " ";
+        if (seekPossible) {
+            cmdline += "-re -ss " + std::string(ms_to_ffmpeg_time(start_at_ms)) + " ";
+        } else {
+            cmdline += "-re ";
+        }
 
         std::string ninput(input_file);
         replaceAll(ninput, "&", "\\&");
@@ -346,6 +357,10 @@ bool TranscodeFFmpeg::fork_ffmpeg(long start_at_ms) {
 }
 
 void TranscodeFFmpeg::seek_video(const char* ms) {
+    if (!seekPossible) {
+        return;
+    }
+
     CONSOLE_TRACE("seek_video, stop video");
     stop_video();
 
@@ -358,11 +373,15 @@ void TranscodeFFmpeg::seek_video(const char* ms) {
 }
 
 void TranscodeFFmpeg::pause_video() {
-    kill(ffmpeg_pid, SIGSTOP);
+    if (ffmpeg_pid > 0) {
+        kill(ffmpeg_pid, SIGSTOP);
+    }
 }
 
 void TranscodeFFmpeg::resume_video() {
-    kill(ffmpeg_pid, SIGCONT);
+    if (ffmpeg_pid > 0) {
+        kill(ffmpeg_pid, SIGCONT);
+    }
 }
 
 void TranscodeFFmpeg::stop_video() {
